+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
 ## [Unreleased]
--  [ENH] Added a `complete` method for polars. - Issue #1352 @samukweku
--  [ENH] Added a `row_to_names` method for polars. Issue #1352
--  [ENH] `read_commandline` function now supports polars - Issue #1352
+
+- [ENH] Added a `complete` method for polars. - Issue #1352 @samukweku
+- [ENH] Added a `row_to_names` method for polars. Issue #1352
+- [ENH] `read_commandline` function now supports polars - Issue #1352
 
 - [ENH] Improved performance for non-equi joins when using numba - @samukweku PR #1341
 - [ENH] Added a `clean_names` method for polars - it can be used to clean the column names, or clean column values . Issue #1343 @samukweku
