+++ janitor/polars/lazyframe.py
@@ -1,10 +1,13 @@
 from __future__ import annotations
 
+from typing import Any
+
 from polars.type_aliases import ColumnNameOrSelector
 
 from janitor.utils import import_message
 
 from .clean_names import _clean_column_names
+from .complete import _complete
 from .pivot_longer import _pivot_longer
 from .row_to_names import _row_to_names
 
@@ -311,4 +314,115 @@ class PolarsLazyFrame:
             remove_rows=remove_rows,
             remove_rows_above=remove_rows_above,
             separator=separator,
+        )
+
+    def complete(
+        self,
+        *columns: ColumnNameOrSelector,
+        fill_value: dict | Any | pl.Expr = None,
+        explicit: bool = True,
+        sort: bool = False,
+        by: ColumnNameOrSelector = None,
+    ) -> pl.LazyFrame:
+        """
+        Turns implicit missing values into explicit missing values.
+
+        It is modeled after tidyr's `complete` function.
+        In a way, it is the inverse of `pl.drop_nulls`,
+        as it exposes implicitly missing rows.
+
+        If the combination involves multiple columns, pass it as a struct,
+        with an alias - the name of the struct should not exist in the LazyFrame.
+
+        If new values need to be introduced, a polars Expression
+        with the new values can be passed, as long as the polars Expression
+        has a name that already exists in the LazyFrame.
+
+        It is up to the user to ensure that the polars expression returns
+        unique values.
+
+        Note that if the polars expression evaluates to a struct,
+        then the fields, not the name, should already exist in the LazyFrame.
+
+        Examples:
+            >>> import polars as pl
+            >>> import janitor.polars
+            >>> df = pl.LazyFrame(
+            ...     dict(
+            ...         group=(1, 2, 1, 2),
+            ...         item_id=(1, 2, 2, 3),
+            ...         item_name=("a", "a", "b", "b"),
+            ...         value1=(1, None, 3, 4),
+            ...         value2=range(4, 8),
+            ...     )
+            ... )
+            >>> df.collect()
+            shape: (4, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Generate all possible combinations of
+            `group`, `item_id`, and `item_name`
+            (whether or not they appear in the data)
+            >>> with pl.Config(tbl_rows=-1):
+            ...     df.janitor.complete("group", "item_id", "item_name", sort=True).collect()
+            shape: (12, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 1       ┆ b         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 3       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+        !!! info "New in version 0.28.0"
+
+        Args:
+            *columns: This refers to the columns to be completed.
+                It can be a string or a column selector or a polars expression.
+                A polars expression can be used to introduced new values,
+                as long as the polars expression has a name that already exists
+                in the LazyFrame.
+                It is up to the user to ensure that the polars expression returns
+                unique values.
+            fill_value: Scalar value or polars expression to use instead of nulls
+                for missing combinations. A dictionary, mapping columns names
+                to a scalar value is also accepted.
+            explicit: Determines if only implicitly missing values
+                should be filled (`False`), or all nulls existing in the LazyFrame
+                (`True`). `explicit` is applicable only
+                if `fill_value` is not `None`.
+            sort: Sort the LazyFrame based on *columns.
+            by: Column(s) to group by.
+                The explicit missing rows are returned per group.
+
+        Returns:
+            A polars LazyFrame.
+        """  # noqa: E501
+        return _complete(
+            df=self._df,
+            columns=columns,
+            fill_value=fill_value,
+            explicit=explicit,
+            sort=sort,
+            by=by,
         )
